+++ src/main/c/graph.h
@@ -3,7 +3,7 @@
 
 struct Graph {
   i32 edges_size;
-  i32* edges;
+  i8* edges;
 };
 
 typedef struct Graph* Graph;
@@ -12,8 +12,8 @@ Graph graph_new(i32);
 
 void graph_free(Graph);
 
-void graph_add_edge(Graph, i32, i32, i32);
+void graph_add_edge(Graph, i32, i32, i8);
 
-i32 graph_get_weight(Graph, i32, i32);
+i8 graph_get_weight(Graph, i32, i32);
 
 List graph_shortest_path(Graph, i32, i32);
